move packetin_test.go code and add Priority filed in GDP entry programming
